Stabilize static rope handoff and rendering
Reworked client rope/static-mesh coordination so sims are explicitly disturbed and woken on terrain edits, chunk payload changes, contact pulses, and tuning/preset updates instead of relying on delayed invalidation. Added section build generations, accepted/claim overlap timing, retiring-mesh grace handling, wind cooldown checks, and static-contact wake logic to prevent stale mesh reuse and reduce handoff popping. Rendering/static geometry now keeps world positions in double precision and introduces bounded Hermite midpoint smoothing so dynamic and baked ropes share shape continuity (including large-coordinate chunk-local conversion fixes). Also added topology variant caching in RopeTuning plus targeted unit coverage for wake logic, generation tracking, smoothing, and new rendering helpers.

+++ src/main/java/com/zhongbai233/super_lead/lead/client/chunk/RopeSectionLine.java
@@ -2,8 +2,8 @@ package com.zhongbai233.super_lead.lead.client.chunk;
 
 /** Static chunk-section line segment used for attachment hanger strings. */
 public record RopeSectionLine(
-        float ax, float ay, float az,
-        float bx, float by, float bz,
+        double ax, double ay, double az,
+        double bx, double by, double bz,
         int color,
         int light) {
 }
